feat: added ability to read assay plate barcodes
Using Pandas to read in an Excel (.xlsx) file that contains a list
of assay plate barcode and then populate a dictionary, indexed
on barcodes, with the Plate object. Each plate obj contains info
about the number of wells and the plate part number.

Also updated the input_arguments.py file to use the appropriate
default values.