feat: Plate objects now hold their own protocols
To make each plate autonomous, I added a method for maintaining
the protocol (refs & instructions) within each plate object.
This way, each plate carries a set of instructions that are
designed as part of the experiment. The ultimate goal is to see
if plate processing can be parallelized once instruction duration
is inculded in the object parameters. To obtain this functionality
it might be necessary to have a function that overlaps each
plate protocol and iterates through the instructions looking for
ways to interleve them and then building a full protocol file
from the individual plate instructions in the right way.